function that calculates the balance of a binary tree

+++ 0x1D-binary_trees/14-binary_tree_balance.c
@@ -7,8 +7,8 @@
  */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	size_t left_size = 0, right_size = 0;
-	size_t end = 0;
+	int left_size = 0, right_size = 0;
+	int end = 0;
 
 	end++;
 	if (!tree)
